Examples/APIBackend/requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.3.3
Beaker==1.10.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.6.1
defusedxml==0.6.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
flask-talisman==0.7.0
future==0.17.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
oic==1.2.0
pycparser==2.19
pycryptodomex==3.7.3
pyjwkest==1.4.0
PyJWT==1.7.1
pylint==2.4.4
pylint-runner==0.5.4
pyOpenSSL==19.0.0
python-dotenv==0.10.3
requests==2.21.0
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4.2
urllib3==1.24.2
Werkzeug==0.15.3
wrapt==1.11.2
